+++ pyfdc/pyfdc.py
@@ -1,38 +1,41 @@
 # Accesses the food search endpoint
 import requests
 import json
-import pandas as pd
+from pandas import DataFrame
 from pandas.io.json import json_normalize
 from itertools import chain
-import os
-from utils import key_signup
+from pyfdc.utils import key_signup
+from sys import exit
+from requests.auth import HTTPBasicAuth
+from bs4 import BeautifulSoup
 
 
-class FoodSearch(object):
+class FoodDataCentral(object):
     """
 
-    This class provides access to and manipulation of the Food Data Central food search end point.
+    This class provides access to and manipulation of the Food Data Central food search and details end points.
 
     For more details, please see: https://fdc.nal.usda.gov/api-guide.html
 
+    This uses version one of the API access point.
+
     """
 
-    def __init__(self, search_phrase, brand_owner=None, ingredients=None):
-
-        if "api_key" in os.environ:
-            self.api_key = os.environ.get("api_key")
+    def __init__(self):
+        if "pyfdc_key" in os.environ:
+            self.api_key = os.environ.get("pyfdc_key")
         else:
             key_signup()
 
-        self.search_phrase = search_phrase
-        self.brand_owner = brand_owner
-        self.ingredients = ingredients
-
-    def get_food_info(self, target=None, page_number=None, require_all=True, sort_field='publishedDate',
-                      sort_direction='asc'):
+    def get_food_info(self, search_phrase=None, ingredients=None, brand_owner= None,
+                      target=None, page_number=None, page_size = 50,
+                      sort_field=None, sort_direction='asc'):
 
         """
 
+        :param brand_owner: str Defaults to None
+        :param ingredients: str to limit the search to certain ingredients
+        :param search_phrase: str A search phrase eg "chicken"
         :param target: A string specifying which of the available values should be returned. Can also
         be a list of strings.
 
@@ -45,74 +48,71 @@ class FoodSearch(object):
         :return: A generator object with the required results.
 
         """
-        search_query = {'generalSearchInput': self.search_phrase,
-                        'ingredients': self.ingredients,
-                        'requireAllWords': require_all,
-                        'pageNumber': page_number,
-                        'sortField': sort_field,
-                        'sortDirection': sort_direction,
-                        'brandOwner': self.brand_owner}
-        # Maybe this can be done with a decorator, I just don't know how(yet)
-        available_targets = ["fdcId", "description", "scientificName", "commonNames",
-                             "additionalDescriptions", "dataType", "foodCode",
-                             "gtinUpc", "ndbNumber", "publishedDate", "brandOwner",
-                             "ingredients", "allHighlightFields", "score"]
 
+        search_query = {'query': search_phrase,
+                        'ingredients': ingredients,
+                        'pageSize': page_size,
+                        'pageNumber': page_number,
+                        'sortBy': sort_field,
+                        'sortOrder': sort_direction,
+                        'brandOwner': brand_owner}
+        # alias camel with snake case
+        available_targets = {"fdc_id": 'fdcId',
+                             "description": 'description',
+                             "scientific_name": 'scientificName',
+                             "common_names": 'commonNames',
+                             "additional_descriptions": 'additionalDescriptions',
+                             "gtin_upc": 'gtinUpc',
+                             "ndb_number": 'ndbNumber',
+                             "published_date": 'publicationDate',
+                             "brand_owner": 'brandOwner',
+                             "ingredients": 'ingredients',
+                             "score": 'score'}
+        headers = {'Accept':'application/json'}
         request_parameters = {'api_key': self.api_key}
-        url_response = requests.post(r"https://api.nal.usda.gov/fdc/v1/search",
-                                     json=search_query, params=request_parameters)
+        # docs
+        # https://fdc.nal.usda.gov/api-spec/fdc_api.html#/FDC/postFoodsSearch
+        url_response = requests.get("https://api.nal.usda.gov/fdc/v1/foods/search?api_key={}".format(self.api_key),
+                                     params = search_query)
         try:
             url_response.raise_for_status()
             unprocessed_result = json.loads(url_response.content)["foods"]
-            if target is None or target not in available_targets:
-                raise ValueError("target should be one of {}".format(available_targets))
+            if target is None or target not in available_targets.keys():
+                raise KeyError("target should be one of {}".format(available_targets.keys()))
 
             else:
                 for x in unprocessed_result:
-                    yield [value for key, value in x.items() if key == target]
+                    yield [value for key, value in x.items() if key == available_targets[target]]
 
         except requests.exceptions.HTTPError as error:
             print(error)
-            sys.exit(1)
-
-    def get_multiple_details(self, target_fields=None):
-        result = []
-        for target_key in target_fields:
-            result.append(list(self.get_food_info(target_key)))
-
-        return pd.DataFrame(list(map(lambda x: list(chain.from_iterable(x)), result)),
-                            index=target_fields).transpose()
 
 
-# Accesses the food details endpoint hence the name
-class FoodDetails(object):
-    """
-
-     This class provides access to and manipulation of the Food Data Central food details end point.
-
-     For more details, please see: https://fdc.nal.usda.gov/api-guide.html
-
-     """
-
-    def __init__(self, fdc_id):
-        # There must be a way to avoid repeating this, I haven't learnt how yet
-
-        if "api_key" in os.environ:
-            self.api_key = os.environ.get("api_key")
-        else:
-            key_signup()
-
-        self.fdc_id = fdc_id
-
-    def get_food_details(self, target_field=None):
+    def get_multiple_details(self, search_phrase=None, target_fields=None, **kwargs):
         """
 
+        :param target_fields: A list of targets eg ['fdc_id','description']
+        :return: A pandas DataFrame
+        """
+        result = []
+        for target_key in target_fields:
+            result.append(list(self.get_food_info(search_phrase=search_phrase, target=target_key,**kwargs)))
+
+        return DataFrame(list(map(lambda x: list(chain.from_iterable(x)), result)),
+                            index=target_fields).transpose()
+
+    def get_food_details(self, fdc_id=None, target_field=None):
+        """
+
+        Accesses the FoodDetails EndPoint
+
+        :param fdc_id: A FoodDataCentral Food ID
         :param target_field: A string indicating which field to return
 
         :return: A JSON object with the desired results.
 
         """
-        base_url = "https://api.nal.usda.gov/fdc/v1/{}?api_key={}".format(self.fdc_id, self.api_key)
+        base_url = "https://api.nal.usda.gov/fdc/v1/{}?api_key={}".format(fdc_id, self.api_key)
         url_response = requests.get(base_url)
         try:
             url_response.raise_for_status()
@@ -122,17 +122,12 @@ class FoodDetails(object):
                 return url_response.json()[target_field]
         except requests.exceptions.HTTPError as error:
             print(error)
-            sys.exit(1)
 
-    def get_nutrients(self):
+
+    def get_nutrients(self, **kwargs):
         """
         :return: A DataFrame showing nutrient details
 
         """
-        use_object = self.get_food_details("foodNutrients")
-        return json_normalize(pd.DataFrame(use_object)["nutrient"])
-
-
-
-
-
+        use_object = self.get_food_details(target_field="foodNutrients",**kwargs)
+        return json_normalize(DataFrame(use_object)["nutrient"])