+++ changelog.md
@@ -1,5 +1,13 @@
 # pyfdc 0.1.3
 
+**Major Changes**
+
+- Classes FoodSearch and FoodDetails have been dropped. Use `FoodDataCentral` instead.
+
+- `api_key` was renamed to `pyfdc_api_key`
+
+- `target_field` in `get_food_info` now uses snake_case instead of CamelCase. 
+
 **Major additions**
 
 - Added an option to signup for an api key